+++ posthog-android/src/main/java/com/posthog/android/PostHogAndroidConfig.kt
@@ -1,6 +1,7 @@
 package com.posthog.android
 
 import com.posthog.PostHogConfig
+import com.posthog.PostHogExperimental
 import com.posthog.android.replay.PostHogSessionReplayConfig
 
 /**
@@ -16,5 +17,6 @@ public open class PostHogAndroidConfig(
     public var captureApplicationLifecycleEvents: Boolean = true,
     public var captureDeepLinks: Boolean = true,
     public var captureScreenViews: Boolean = true,
+    @PostHogExperimental
     public var sessionReplayConfig: PostHogSessionReplayConfig = PostHogSessionReplayConfig(),
 ) : PostHogConfig(apiKey, host)