+++ posthog-android/src/main/java/com/posthog/android/replay/internal/ViewTreeSnapshotStatus.kt
@@ -4,7 +4,7 @@ import com.posthog.internal.replay.RRWireframe
 
 internal class ViewTreeSnapshotStatus(
     val listener: NextDrawListener,
-//    var sentFullSnapshot: Boolean = false,
+    var sentFullSnapshot: Boolean = false,
     var sentMetaEvent: Boolean = false,
     var lastSnapshot: RRWireframe? = null,
 )