+++ posthog/src/main/java/com/posthog/PostHogExperimental.kt
@@ -1,0 +1,17 @@
+package com.posthog
+
+/**
+ * Any target annotated with this annotation is an experimental feature
+ */
+@Retention(AnnotationRetention.BINARY)
+@Target(
+    AnnotationTarget.ANNOTATION_CLASS,
+    AnnotationTarget.CLASS,
+    AnnotationTarget.FUNCTION,
+    AnnotationTarget.PROPERTY_GETTER,
+    AnnotationTarget.PROPERTY_SETTER,
+    AnnotationTarget.CONSTRUCTOR,
+    AnnotationTarget.FIELD,
+    AnnotationTarget.FILE,
+)
+public annotation class PostHogExperimental