+++ posthog/src/main/java/com/posthog/internal/replay/RRIncrementalMutationData.kt
@@ -4,8 +4,9 @@ import com.posthog.PostHogInternal
 
 @PostHogInternal
 public class RRIncrementalMutationData(
-    public val adds: List<RRAddedNode>? = null,
+    public val adds: List<RRMutatedNode>? = null,
     public val removes: List<RRRemovedNode>? = null,
+    // updates and adds share the same format
+    public val updates: List<RRMutatedNode>? = null,
     public val source: RRIncrementalSource = RRIncrementalSource.Mutation,
-    // TODO: mutations/updates pending
 )