+++ posthog/src/main/java/com/posthog/internal/replay/RRMutatedNode.kt
@@ -3,7 +3,7 @@ package com.posthog.internal.replay
 import com.posthog.PostHogInternal
 
 @PostHogInternal
-public class RRAddedNode(
+public class RRMutatedNode(
     public val wireframe: RRWireframe,
     public val parentId: Int? = null,
 )